requirements.txt
django-bootstrap-form
pandas
sankey
pysankey